Initial Rendering setup done

+++ src/main/java/org/game/jcengine/util/Point.java
@@ -3,15 +3,15 @@ package org.game.jcengine.util;
 import java.util.Objects;
 
 public class Point {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
@@ -19,7 +19,7 @@ public class Point {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
